+++ packages/zend-captcha/library/Zend/Captcha/Image.php
@@ -581,8 +581,10 @@ class Zend_Captcha_Image extends Zend_Captcha_Word
         }
 
         imagepng($img2, $img_file);
-        imagedestroy($img);
-        imagedestroy($img2);
+        if (PHP_VERSION_ID < 80000) {
+            imagedestroy($img);
+            imagedestroy($img2);
+        }
     }
 
     /**
@@ -616,7 +618,7 @@ class Zend_Captcha_Image extends Zend_Captcha_Word
      * @param mixed $element
      * @return string
      */
-    public function render(Zend_View_Interface $view = null, $element = null)
+    public function render(?Zend_View_Interface $view = null, $element = null)
     {
         $endTag = ' />';
         if (($view instanceof Zend_View_Abstract) && !$view->doctype()->isXhtml()) {