+++ packages/zend-controller/library/Zend/Controller/Front.php
@@ -287,7 +287,7 @@ class Zend_Controller_Front
     {
         try{
             $dir = new DirectoryIterator($path);
-        } catch(Exception $e) {
+        } catch(Throwable $e) {
             // require_once 'Zend/Controller/Exception.php';
             throw new Zend_Controller_Exception("Directory $path not readable", 0, $e);
         }
@@ -832,7 +832,7 @@ class Zend_Controller_Front
      * @param Zend_Controller_Response_Abstract|null $response
      * @return void|Zend_Controller_Response_Abstract Returns response object if returnResponse() is true
      */
-    public function dispatch(Zend_Controller_Request_Abstract $request = null, Zend_Controller_Response_Abstract $response = null)
+    public function dispatch(?Zend_Controller_Request_Abstract $request = null, ?Zend_Controller_Response_Abstract $response = null)
     {
         if (!$this->getParam('noErrorHandler') && !$this->_plugins->hasPlugin('Zend_Controller_Plugin_ErrorHandler')) {
             // Register with stack index of 100
@@ -909,11 +909,13 @@ class Zend_Controller_Front
 
             try {
                 $router->route($this->_request);
-            }  catch (Exception $e) {
+            }  catch (Throwable $e) {
                 if ($this->throwExceptions()) {
                     throw $e;
                 }
-
+                if (!$e instanceof Exception) {
+                    $e = new Zend_Controller_Exception($e->getMessage(), $e->getCode(), $e);
+                }
                 $this->_response->setException($e);
             }
 
@@ -952,9 +954,12 @@ class Zend_Controller_Front
                  */
                 try {
                     $dispatcher->dispatch($this->_request, $this->_response);
-                } catch (Exception $e) {
+                } catch (Throwable $e) {
                     if ($this->throwExceptions()) {
                         throw $e;
+                    }
+                    if (!$e instanceof Exception) {
+                        $e = new Zend_Controller_Exception($e->getMessage(), $e->getCode(), $e);
                     }
                     $this->_response->setException($e);
                 }
@@ -964,11 +969,13 @@ class Zend_Controller_Front
                  */
                 $this->_plugins->postDispatch($this->_request);
             } while (!$this->_request->isDispatched());
-        } catch (Exception $e) {
+        } catch (Throwable $e) {
             if ($this->throwExceptions()) {
                 throw $e;
             }
-
+            if (!$e instanceof Exception) {
+                $e = new Zend_Controller_Exception($e->getMessage(), $e->getCode(), $e);
+            }
             $this->_response->setException($e);
         }
 
@@ -977,11 +984,13 @@ class Zend_Controller_Front
          */
         try {
             $this->_plugins->dispatchLoopShutdown();
-        } catch (Exception $e) {
+        } catch (Throwable $e) {
             if ($this->throwExceptions()) {
                 throw $e;
             }
-
+            if (!$e instanceof Exception) {
+                $e = new Zend_Controller_Exception($e->getMessage(), $e->getCode(), $e);
+            }
             $this->_response->setException($e);
         }