+++ packages/zend-locale/library/Zend/Locale/Data.php
@@ -1586,7 +1586,7 @@ class Zend_Locale_Data
      */
     public static function disableCache($flag)
     {
-        self::$_cacheDisabled = (boolean) $flag;
+        self::$_cacheDisabled = (bool) $flag;
     }
 
     /**