+++ packages/zend-rest/library/Zend/Rest/Server.php
@@ -266,7 +266,7 @@ class Zend_Rest_Server implements Zend_Server_Interface
                                 $this->_functions[$this->_method]->getName(),
                                 $callingArgs
                             );
-                        } catch (Exception $e) {
+                        } catch (Throwable $e) {
                             $result = $this->fault($e);
                         }
                     }
@@ -478,8 +478,9 @@ class Zend_Rest_Server implements Zend_Server_Interface
      */
     public function fault($exception = null, $code = null)
     {
-        if (isset($this->_functions[$this->_method])) {
-            $function = $this->_functions[$this->_method];
+        $method = $this->_method ?? '';
+        if (isset($this->_functions[$method])) {
+            $function = $this->_functions[$method];
         } elseif (isset($this->_method)) {
             $function = $this->_method;
         } else {
@@ -514,7 +515,7 @@ class Zend_Rest_Server implements Zend_Server_Interface
         $xmlResponse = $dom->createElement('response');
         $xmlMethod->appendChild($xmlResponse);
 
-        if ($exception instanceof Exception) {
+        if ($exception instanceof Throwable) {
             $element = $dom->createElement('message');
             $element->appendChild(
                 $dom->createTextNode($exception->getMessage())
@@ -629,7 +630,7 @@ class Zend_Rest_Server implements Zend_Server_Interface
                 ),
                 $args
             );
-        } catch (Exception $e) {
+        } catch (Throwable $e) {
             $result = $this->fault($e);
         }
         return $result;
@@ -651,7 +652,7 @@ class Zend_Rest_Server implements Zend_Server_Interface
             } else {
                 $object = $this->_functions[$this->_method]->getDeclaringClass()->newInstance();
             }
-        } catch (Exception $e) {
+        } catch (Throwable $e) {
             // require_once 'Zend/Rest/Server/Exception.php';
             throw new Zend_Rest_Server_Exception(
                 'Error instantiating class ' . $class .
@@ -668,7 +669,7 @@ class Zend_Rest_Server implements Zend_Server_Interface
                 $object,
                 $args
             );
-        } catch (Exception $e) {
+        } catch (Throwable $e) {
             $result = $this->fault($e);
         }